+++ packages/components/src/table/index.tsx
@@ -3,7 +3,7 @@ import styled from "styled-components";
 import { GlobalStyles } from "../app";
 import Thead from "./table-head";
 import TableBody from "./table-body";
-import Pagination from "../pagination";
+import Pagination, { PaginationCallbackProps } from "../pagination";
 import LoadingMask from "../loading/loading-mask";
 import Col from "../grid/col";
 import Row from "../grid/row";
@@ -42,6 +42,7 @@ interface TableProps {
 	dataSource: any[];
 	/** Columns of table */
 	columns: any[];
+	/** callback to be executed when any row is clicked */
 	onRow?: (record: any, rowIndex: string) => OnRowReturn;
 	/** Config of pagination. You can ref pagination document
 	 * at https://mint-ui.netlify.com/?path=/story/pagination--basic
@@ -50,10 +51,19 @@ interface TableProps {
 	pagination?: Pagination | boolean;
 	/**	Loading status of table */
 	loading?: boolean;
-	/** Callback executed when pagination is changed */
-	onChange?: (pagination: any) => void;
+	/** Callback executed when pagination, sorting or filters are changed */
+	onChange?: (
+		pagination: any,
+		sorter: any,
+		filters?: any,
+		extra?: any
+	) => void;
 }
 
+export interface ActiveSort {
+	column: ColumnType;
+	order: sortOrderType;
+}
 export const Table = ({
 	dataSource = [],
 	columns,
@@ -64,31 +74,35 @@ export const Table = ({
 }: TableProps) => {
 	const [data, setData] = React.useState(dataSource);
 	const defaultCol = columns.filter(column => column.defaultSortOrder)[0];
-	const [activeCol, setActiveCol] = React.useState<ColumnType>(defaultCol);
-	const [currentPage, setCurrentPage] = React.useState<number>(
+	const [activeSort, setActiveSort] = React.useState<ActiveSort>({
+		column: defaultCol,
+		order: "ascends"
+	});
+	const [activePagination, setActivePagination] = React.useState<
+		PaginationCallbackProps
+	>(
 		pagination && typeof pagination !== "boolean"
-			? pagination.currentPage
-			: 1
-	);
-	const [pageSize] = React.useState<number>(
-		pagination && typeof pagination !== "boolean" ? pagination.pageSize : 10
+			? {
+					currentPage: pagination.currentPage,
+					pageSize: pagination.pageSize
+			  }
+			: { currentPage: 1, pageSize: 10 }
 	);
 	React.useEffect(() => {
 		setData(dataSource);
 	}, [dataSource]);
 
 	React.useEffect(() => {
-		const pagination = { currentPage, pageSize };
-		onChange ? onChange(pagination) : null;
-	}, [currentPage, pageSize]);
+		onChange ? onChange(activePagination, activeSort) : null;
+	}, [activePagination, activeSort]);
 
 	React.useEffect(() => {
-		handleSort(activeCol, activeCol.defaultSortOrder);
+		handleSort(activeSort.column, activeSort.column.defaultSortOrder);
 	}, []);
 
 	const handleSort = (
 		column: ColumnType,
-		sortOrder: sortOrderType | undefined
+		sortOrder: sortOrderType = "ascends"
 	) => {
 		const sortedData = data.sort((a: any, b: any) => {
 			if (column && column.sorter) {
@@ -97,7 +111,7 @@ export const Table = ({
 			return column;
 		});
 		setData([...sortedData]);
-		setActiveCol(column);
+		setActiveSort({ column, order: sortOrder });
 	};
 
 	return (
@@ -106,42 +120,38 @@ export const Table = ({
 				{loading ? <LoadingMask /> : null}
 				<Thead
 					columns={columns}
-					activeCol={activeCol}
+					activeSort={activeSort}
 					handleSort={handleSort}
 				/>
 				<TableBody
 					data={data}
 					onRow={onRow}
-					pageSize={pageSize}
-					currentPage={currentPage}
+					pageSize={activePagination.pageSize}
+					currentPage={activePagination.currentPage}
 					isAjax={pagination ? true : false}
 				/>
 			</TableEl>
 			{data.length ? (
 				<Row style={{ marginTop: "20px" }}>
 					<Col>
-						{typeof pagination === "boolean" &&
-						pagination === false ? null : (
+						{!pagination ? null : (
 							<Pagination
-								style={{
-									display: "flex",
-									justifyContent: "flex-end"
-								}}
 								total={
-									pagination &&
 									typeof pagination !== "boolean"
 										? pagination.total
 										: dataSource.length
 								}
-								current={
-									pagination &&
-									typeof pagination !== "boolean"
-										? pagination.currentPage
-										: currentPage
-								}
-								onChange={(page: number) =>
-									setCurrentPage(page)
-								}
+								current={activePagination.currentPage}
+								pageSize={activePagination.pageSize}
+								onChange={(
+									paginationObj: PaginationCallbackProps
+								) => {
+									setActivePagination(paginationObj);
+								}}
+								style={{
+									display: "flex",
+									justifyContent: "flex-end"
+								}}
 							/>
 						)}
 					</Col>